+++ packages/kepler/src/components/KeplerAddDataDialog.tsx
@@ -91,9 +91,9 @@ export const KeplerAddDataDialog = ({
   loadFiles,
   listS3Files,
   loadS3Files,
-  saveS3Credential,
+  saveS3Credentials,
   loadS3Credentials,
-  deleteS3Credential,
+  deleteS3Credentials,
   s3Browser,
   method = AddDataMethods.Upload,
   acceptedFormats = DEFAULT_ACCEPTED_FORMATS,
@@ -122,9 +122,7 @@ export const KeplerAddDataDialog = ({
         <DialogContent className="min-w-md h-[80vh] max-w-4xl grid-rows-[auto,1fr]">
           <DialogHeader>
             <DialogTitle>Add Data</DialogTitle>
-            <DialogDescription>
-              Add data to the project.
-            </DialogDescription>
+            <DialogDescription>Add data to the project.</DialogDescription>
           </DialogHeader>
           <Tabs
             defaultValue={currentMethod}
@@ -187,9 +185,9 @@ export const KeplerAddDataDialog = ({
                 listS3Files={listS3Files}
                 s3Browser={s3Browser}
                 loadS3Files={onLoadS3Files}
-                saveS3Credential={saveS3Credential}
+                saveS3Credentials={saveS3Credentials}
                 loadS3Credentials={loadS3Credentials}
-                deleteS3Credential={deleteS3Credential}
+                deleteS3Credentials={deleteS3Credentials}
               />
             </TabsContent>
           </Tabs>