+++ packages/s3-browser/src/index.ts
@@ -3,6 +3,6 @@
  * @packageDocumentation
  */
 export {default as S3FileBrowser} from './S3FileBrowser';
-export {S3CredentialForm} from './S3CredentialForm';
+export {S3CredentialsForm} from './S3CredentialsForm';
 export {createS3BrowserSlice, type S3BrowserState} from './S3BrowserSlice';
 export * from '@sqlrooms/s3-browser-config';